backend/mcp-server/requirements.txt
# NYC POI Concierge MCP Server Dependencies

# Model Context Protocol
mcp>=0.9.0

# Database
pymongo>=4.6.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.5.0

# External APIs
tavily-python>=0.3.6
